// ionic-app/src/app/models/interfaces.ts
export interface IRecord<T> {
	id: string;
	data?: T;
}

export interface IUserTeamInfo {
	title: string;
	retroItems?: { [type: string]: IRetroItem[] };
}

export interface IUserTeamInfoWithId extends IUserTeamInfo {
	id: string;
}

export interface IUser {
	title: string;
	email?: string;
	emailVerified?: boolean;
	avatar?: IAvatar;
	teams?: { [id: string]: IUserTeamInfo };
	dataTugProjects?: IDataTugProjectBrief[];
}

export interface IDataTugProjectBrief {
	id: string;
	title: string;
	userTitle?: string;
}
// ...
export interface IAvatar {
	gravatar?: string;
	external?: {
		provider?: string;
		url?: string;
	};
}
// ...
export interface IRetroItem {
	// eslint-disable-next-line @typescript-eslint/naming-convention
	ID: string; // see comments on RetroItem in Go why it's not `id` in all lower case
	type?: string; // provided only for root level items
	title: string;
	children?: IRetroItem[];
	likedUserIds?: string[];
}